requirements.txt
# Web framework
Flask==3.1.2

# PyTorch stack
torch==2.2.0
torchvision==0.17.1
torchaudio==2.2.1

# OCR / image handling
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.26.0

# YOLO from Ultralytics
ultralytics==8.3.243
